refactor: fix name of file - transation -> transaction-schema.ts
